Move structure detection tests to existing file
