+++ tests/tests/test_unit/test_detect/test_filters/test_volume_filters/test_structure_detection.py
@@ -1,17 +1,16 @@
 import numpy as np
 import pytest
 
-from cellfinder_core.detect.filters.volume import structure_detection
+from cellfinder_core.detect.filters.volume.structure_detection import (
+    CellDetector,
+    get_non_zero_ull_min_wrapper,
+    get_structure_centre_wrapper,
+)
 
 
 def test_get_non_zero_ull_min():
-    assert (
-        structure_detection.get_non_zero_ull_min_wrapper(list(range(10))) == 1
-    )
-    assert (
-        structure_detection.get_non_zero_ull_min_wrapper([0] * 10)
-        == (2**64) - 1
-    )
+    assert get_non_zero_ull_min_wrapper(list(range(10))) == 1
+    assert get_non_zero_ull_min_wrapper([0] * 10) == (2**64) - 1
 
 
 class Point:
@@ -43,9 +42,73 @@ def structure(three_d_cross):
 
 
 def test_get_structure_centre(structure):
-    result_point = structure_detection.get_structure_centre_wrapper(structure)
+    result_point = get_structure_centre_wrapper(structure)
     assert (result_point["x"], result_point["y"], result_point["z"]) == (
         1,
         1,
         1,
     )
+
+
+# Image volume dimensions
+width, height = 3, 2
+depth = 2
+
+# Each item in the test data contains:
+#
+# - A list of indices to mark as structure pixels
+# - A dict of expected structure coordinates
+test_data = [
+    (
+        # No pixels
+        [],
+        {},
+    ),
+    (
+        # Two pixels connected in a single structure along x
+        [(0, 0, 0), (0, 1, 0)],
+        {1: [{"x": 0, "y": 0, "z": 0}, {"x": 1, "y": 0, "z": 0}]},
+    ),
+    (
+        # Two pixels connected in a single structure along y
+        [(0, 0, 0), (0, 0, 1)],
+        {1: [{"x": 0, "y": 0, "z": 0}, {"x": 0, "y": 1, "z": 0}]},
+    ),
+    (
+        # Two pixels connected in a single structure along z
+        [(0, 0, 0), (1, 0, 0)],
+        {1: [{"x": 0, "y": 0, "z": 0}, {"x": 0, "y": 0, "z": 1}]},
+    ),
+    (
+        # Four pixels all connected and spread across x-y-z
+        [(0, 0, 0), (1, 0, 0), (1, 1, 0), (1, 0, 1)],
+        {
+            1: [
+                {"x": 0, "y": 0, "z": 0},
+                {"x": 0, "y": 0, "z": 1},
+                {"x": 1, "y": 0, "z": 1},
+                {"x": 0, "y": 1, "z": 1},
+            ]
+        },
+    ),
+    (
+        # Two disconnected single-pixel structures
+        [(0, 0, 0), (0, 2, 0)],
+        {1: [{"x": 0, "y": 0, "z": 0}], 2: [{"x": 2, "y": 0, "z": 0}]},
+    ),
+]
+
+
+@pytest.mark.parametrize("pixels,expected_coords", test_data)
+def test_detection(pixels, expected_coords):
+    data = np.zeros((depth, width, height)).astype(np.uint64)
+    detector = CellDetector(width, height, start_z=0)
+
+    for pix in pixels:
+        data[pix] = detector.SOMA_CENTRE_VALUE
+
+    for plane in data:
+        detector.process(plane)
+
+    coords = detector.get_coords_list()
+    assert coords == expected_coords
